urlaub-an-der-stiefelspitze/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-crispy-forms==1.7.2
gevent==1.4.0
gunicorn==19.9.0
idna==2.8
mysqlclient==1.4.2.post1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
